Add labels to images
Added missing files for local cloud autotests setup
CLOUD-1256: Return repeater selector for systems in systems list selector
CLOUD-1256: fixed systems list set. Except when there's only one system for the user
CLOUD-1256: pick correct system from the list
CLOUD-1274: CMS - HTML Data Structure - editor configuration
CLOUD-1283: Preview and publish - speed-up
CLOUD-1301: Add instruction block on main admin page
CLOUD-1303: Limit permissions to edit various pages or records
CLOUD-1304: Add human-readable labels for content fields
CLOUD-1367: Read customized settings from DB, not from config file
CLOUD-1372: Speed up cloud portal build relying on CMS
CLOUD-1375: Global contexts
CLOUD-1377: Clean html attributes in WYSIWYG
CLOUD-1378: Incremental content publish
CLOUD-1379: Restructure cms_structure.json
CLOUD-1381: Change in supported languages -> force update
CLOUD-1382: CMS - content manager can see revisions from other customizations
CLOUD-1383: CMS - DataStructure - file
CLOUD-1385: Migrate DATA_TYPES to Choices
CLOUD-1386: Visualize version state in admin
CLOUD-1388: Notifications service (and tasks) - cache refactoring
CLOUD-1389: Make Pages and Revisions editable
CLOUD-1390: Inter-process communication when CMS is updating
CLOUD-1392: Refactor get_custom_config in email_engine
CLOUD-1393: know issues text change
CLOUD-1394: Support renaming for cms_structure
CLOUD-1395: Support order for data_structure
CLOUD-1396: Speed up CMS: track depending files
CLOUD-1397: Fill mail_from in CMS for all customizations before releasing prod
CLOUD-1399: Support hidden contexts
CLOUD-1400: Script for generating cms_structure.json based on directory
CLOUD-1401: Refactor - remove set_latest_customization_version_id, use models.customization_cache instead
CLOUD-1407: Initiate cms with zip archive
CLOUD-1410: Export zip from CMS
CLOUD-1413: CMS - product workflow
CLOUD-1414: CMS- product workflow refinement
CLOUD-1415: Email engine caches - force filldata and update when version_id is changed
CLOUD-1416: Get rid of db_cache
CLOUD-1418: CMS - load contexts to database and update them when needed
CLOUD-1419: Read files for emails directly from CMS
CLOUD-1422: CMS - Product management refinements
Cloud autotests: running almost seamlessly on local cloud
Cloud tests: fixed failures in system spec
Cloud tests: solved most of errors in system page suit
Cloud-portal tests: intermediate changes
Cloud-portal tests: intermediate changes for webadmin
Cloud-portal tests: intermediate changes for webadmin Gruntfile
Cloud-tests: fixed timeouts
Renamed branch vms_3.1_portal_test to cloud_tests
VMS-7091 Add version file to installed server directory
close broken branch
collapse tag fix
fix branch
hanwha. Hew Email for portal. Again

<==== Mercurial data ====>
Branch: vms_3.2_dev

+++ cloud_portal/build_scripts/build_customization.sh
@@ -19,7 +19,7 @@ dir=../customizations/$CUSTOMIZATION/
 
     echo "------------------------------"
     echo "Copy config"
-    cp $dir/cloud_portal.yaml $TARGET_DIR/$CUSTOMIZATION/
+    cp ../cloud/cloud_portal.yaml $TARGET_DIR/$CUSTOMIZATION/
 
 
     echo "------------------------------"
@@ -37,10 +37,13 @@ dir=../customizations/$CUSTOMIZATION/
 
     cp -rf $dir/front_end/styles/* ../front_end/app/styles/custom
 
-    echo "Move fonts"
-    rm -rf $TARGET_DIR/common || true
-    mkdir -p $TARGET_DIR/common/static
-    mv ../front_end/dist/fonts $TARGET_DIR/common/static/fonts
+    if [ "$CUSTOMIZATION" = "default" ]
+    then
+        echo "Move fonts  - only for default"
+        rm -rf $TARGET_DIR/common || true
+        mkdir -p $TARGET_DIR/common/static
+        mv ../front_end/dist/fonts $TARGET_DIR/common/static/fonts
+    fi
 
     echo "Move front_end to destination"
     mv ../front_end/dist $TARGET_DIR/$CUSTOMIZATION/source/static
@@ -114,26 +117,22 @@ dir=../customizations/$CUSTOMIZATION/
         echo "Overwrite them with localized sources"
         cp -rf $lang_dir/web_common/views/* $TARGET_DIR/$CUSTOMIZATION/source/static/lang_$LANG/web_common/views/ || true
 
+        echo "Generate language.json"
+        pushd $TARGET_DIR/$CUSTOMIZATION/source
+        python ../../../../build_scripts/generate_language_json.py $LANG
+        popd
+
     done
-    rm -rf $TARGET_DIR/$CUSTOMIZATION/source/static/views
 
-    echo "Localization success"
-
-
-    echo "------------------------------"
-    echo "Branding"
-    cp $dir/branding.ts $TARGET_DIR/$CUSTOMIZATION
     pushd $TARGET_DIR/$CUSTOMIZATION/source
-    python ../../../../build_scripts/generate_languages_json.py
-    # python ../../../../build_scripts/branding.py # do not do branding here, leave it to cms
-    rm -rf *.ts
+    python ../../../../build_scripts/generate_all_languages_json.py
     popd
-    echo "Branding success"
 
+    rm -rf $TARGET_DIR/$CUSTOMIZATION/source/static/views
+    echo "Localization success"
 
     echo "copy sources to root"
     cp -rf $TARGET_DIR/$CUSTOMIZATION/source/* $TARGET_DIR/$CUSTOMIZATION/
-
 
 echo "$CUSTOMIZATION Done"
